Move PsbtSighashType to its own module
In preparation for teasing the serialization stuff apart move the
sighash type to its own module.

This is technically a breaking change because we do not re-export the
type from `map::input`. We do maintain the re-export however from
`psbt`.
