+++ bitcoin/src/psbt/map/input.rs
@@ -1,15 +1,12 @@
 // SPDX-License-Identifier: CC0-1.0
 
-use core::fmt;
-use core::str::FromStr;
-
 use hashes::{hash160, ripemd160, sha256, sha256d};
 use secp256k1::XOnlyPublicKey;
 
 use crate::bip32::KeySource;
 use crate::crypto::key::PublicKey;
 use crate::crypto::{ecdsa, taproot};
-use crate::prelude::{btree_map, BTreeMap, Borrow, Box, ToOwned, Vec};
+use crate::prelude::{btree_map, BTreeMap, Borrow, Box, Vec};
 use crate::psbt::consts::{
     PSBT_IN_BIP32_DERIVATION, PSBT_IN_FINAL_SCRIPTSIG, PSBT_IN_FINAL_SCRIPTWITNESS,
     PSBT_IN_HASH160, PSBT_IN_HASH256, PSBT_IN_NON_WITNESS_UTXO, PSBT_IN_PARTIAL_SIG,
@@ -20,10 +17,10 @@ use crate::psbt::consts::{
 };
 use crate::psbt::map::Map;
 use crate::psbt::serialize::Deserialize;
-use crate::psbt::{error, raw, Error};
+use crate::psbt::{error, raw, Error, PsbtSighashType};
 use crate::script::ScriptBuf;
 use crate::sighash::{
-    EcdsaSighashType, InvalidSighashTypeError, NonStandardSighashTypeError, SighashTypeParseError,
+    EcdsaSighashType, InvalidSighashTypeError, NonStandardSighashTypeError,
     TapSighashType,
 };
 use crate::taproot::{ControlBlock, LeafVersion, TapLeafHash, TapNodeHash};
@@ -96,116 +93,6 @@ pub struct Input {
     /// Unknown key-value pairs for this input.
     #[cfg_attr(feature = "serde", serde(with = "crate::serde_utils::btreemap_as_seq_byte_values"))]
     pub unknown: BTreeMap<raw::Key, Vec<u8>>,
-}
-
-/// A Signature hash type for the corresponding input.
-///
-/// As of Taproot upgrade, the signature hash type can be either [`EcdsaSighashType`] or
-/// [`TapSighashType`] but it is not possible to know directly which signature hash type the user is
-/// dealing with. Therefore, the user is responsible for converting to/from [`PsbtSighashType`]
-/// from/to the desired signature hash type they need.
-///
-/// # Examples
-///
-/// ```
-/// use bitcoin::{EcdsaSighashType, TapSighashType};
-/// use bitcoin::psbt::PsbtSighashType;
-///
-/// let _ecdsa_sighash_all: PsbtSighashType = EcdsaSighashType::All.into();
-/// let _tap_sighash_all: PsbtSighashType = TapSighashType::All.into();
-/// ```
-#[derive(Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-pub struct PsbtSighashType {
-    pub(in crate::psbt) inner: u32,
-}
-
-impl fmt::Display for PsbtSighashType {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self.taproot_hash_ty() {
-            Err(_) => write!(f, "{:#x}", self.inner),
-            Ok(taproot_hash_ty) => fmt::Display::fmt(&taproot_hash_ty, f),
-        }
-    }
-}
-
-impl FromStr for PsbtSighashType {
-    type Err = SighashTypeParseError;
-
-    #[inline]
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        // We accept strings of form: "SIGHASH_ALL" etc.
-        //
-        // NB: some of Taproot sighash types are non-standard for pre-Taproot
-        // inputs. We also do not support SIGHASH_RESERVED in verbatim form
-        // ("0xFF" string should be used instead).
-        if let Ok(ty) = s.parse::<TapSighashType>() {
-            return Ok(ty.into());
-        }
-
-        // We accept non-standard sighash values.
-        if let Ok(inner) = u32::from_str_radix(s.trim_start_matches("0x"), 16) {
-            return Ok(PsbtSighashType { inner });
-        }
-
-        Err(SighashTypeParseError { unrecognized: s.to_owned() })
-    }
-}
-impl From<EcdsaSighashType> for PsbtSighashType {
-    fn from(ecdsa_hash_ty: EcdsaSighashType) -> Self {
-        PsbtSighashType { inner: ecdsa_hash_ty as u32 }
-    }
-}
-
-impl From<TapSighashType> for PsbtSighashType {
-    fn from(taproot_hash_ty: TapSighashType) -> Self {
-        PsbtSighashType { inner: taproot_hash_ty as u32 }
-    }
-}
-
-impl PsbtSighashType {
-    /// Ambiguous `ALL` sighash type, may refer to either [`EcdsaSighashType::All`]
-    /// or [`TapSighashType::All`].
-    ///
-    /// This is equivalent to either `EcdsaSighashType::All.into()` or `TapSighashType::All.into()`.
-    /// For sighash types other than `ALL` use the ECDSA or Taproot sighash type directly.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use bitcoin::{EcdsaSighashType, TapSighashType};
-    /// use bitcoin::psbt::PsbtSighashType;
-    /// let _ecdsa_sighash_anyone_can_pay: PsbtSighashType = EcdsaSighashType::AllPlusAnyoneCanPay.into();
-    /// let _tap_sighash_anyone_can_pay: PsbtSighashType = TapSighashType::AllPlusAnyoneCanPay.into();
-    /// ```
-    pub const ALL: PsbtSighashType = PsbtSighashType { inner: 0x01 };
-
-    /// Returns the [`EcdsaSighashType`] if the [`PsbtSighashType`] can be
-    /// converted to one.
-    pub fn ecdsa_hash_ty(self) -> Result<EcdsaSighashType, NonStandardSighashTypeError> {
-        EcdsaSighashType::from_standard(self.inner)
-    }
-
-    /// Returns the [`TapSighashType`] if the [`PsbtSighashType`] can be
-    /// converted to one.
-    pub fn taproot_hash_ty(self) -> Result<TapSighashType, InvalidSighashTypeError> {
-        if self.inner > 0xffu32 {
-            Err(InvalidSighashTypeError(self.inner))
-        } else {
-            TapSighashType::from_consensus_u8(self.inner as u8)
-        }
-    }
-
-    /// Creates a [`PsbtSighashType`] from a raw `u32`.
-    ///
-    /// Allows construction of a non-standard or non-valid sighash flag
-    /// ([`EcdsaSighashType`], [`TapSighashType`] respectively).
-    pub fn from_u32(n: u32) -> PsbtSighashType { PsbtSighashType { inner: n } }
-
-    /// Converts [`PsbtSighashType`] to a raw `u32` sighash flag.
-    ///
-    /// No guarantees are made as to the standardness or validity of the returned value.
-    pub fn to_u32(self) -> u32 { self.inner }
 }
 
 impl Input {
