+++ bitcoin/src/psbt/serialize.rs
@@ -9,7 +9,8 @@ use hashes::{hash160, ripemd160, sha256, sha256d};
 use internals::compact_size;
 use secp256k1::XOnlyPublicKey;
 
-use super::map::{Input, Map, Output, PsbtSighashType};
+use super::PsbtSighashType;
+use super::map::{Input, Map, Output};
 use crate::bip32::{ChildNumber, Fingerprint, KeySource};
 use crate::consensus::encode::{self, deserialize_partial, serialize, Decodable, Encodable};
 use crate::crypto::key::PublicKey;
